lays ground work for messages module and changes
