+++ lib/patient_zero/message.rb
@@ -1,15 +1,16 @@
-module PatientZero
-  class Message < Base
-    def initialize data
-      @data = data
-    end
+require 'patient_zero/message/base'
+require 'patient_zero/message/facebook'
+require 'patient_zero/message/twitter'
+require 'patient_zero/message/instagram'
 
-    def method_missing name, *args
-      if @data.has_key? name.to_s
-        @data.fetch name.to_s
-      else
-        super
-      end
+module PatientZero
+  module Message
+    SOURCE_TYPES = {'TW' => Twitter,
+                    'FB' => Facebook,
+                    'IG' => Instagram}
+
+    def self.for_platform platform, params={}
+      SOURCE_TYPES[platform].new params
     end
   end
 end
