+++ lib/patient_zero/message/facebook.rb
@@ -1,0 +1,13 @@
+module PatientZero
+  module Message
+    class Facebook < Base
+      def engagements
+        @engagements ||= likes + comments + shares + clicks
+      end
+
+      def reach
+
+      end
+    end
+  end
+end
