+++ lib/patient_zero/message/instagram.rb
@@ -1,0 +1,11 @@
+module PatientZero
+  module Message
+    class Instagram < Base
+      def engagements
+      end
+
+      def reach
+      end
+    end
+  end
+end
