+++ lib/patient_zero/message/twitter.rb
@@ -1,0 +1,11 @@
+module PatientZero
+  module Message
+    class Twitter < Base
+      def engagements
+      end
+
+      def reach
+      end
+    end
+  end
+end
